refactor(scripts): extract ghExec into shared gh-exec.js module
Move ghExec utility from follow-up-pr.js into a reusable gh-exec.js module to support code reuse across scripts. Maintains identical logic and error handling for CLI command execution.
